+++ alltheutils/exceptions.py
@@ -1,7 +1,12 @@
+import functools
+import warnings
 from typing import Any, Optional
 
 from alltheutils import TW
-from alltheutils.base_exceptions import custom_exception, custom_exception_str
+from alltheutils.base_exceptions import (
+    CustomBaseException,
+    custom_exception,
+)
 from alltheutils.types import Kwargs
 
 """
@@ -9,21 +14,55 @@ from alltheutils.types import Kwargs
 """
 
 
+# ============================= Top dependencies ===============================
+def deprecated(
+    version: str,
+    replacement: Optional[str] = None,
+    reason: Optional[str] = None,
+):
+    """
+    Decorator to mark functions as deprecated.
+
+    Args:
+        version (str): The version in which the function will be removed.
+        replacement (str, optional): The new function to use instead.
+
+    """
+
+    def decorator(func):
+        func_name = func.__name__  # Get function name automatically
+
+        @functools.wraps(func)
+        def wrapper(*args, **kwargs):
+            message = (
+                f"{func_name}() is deprecated and will be removed in version {version}."
+            )
+            if replacement:
+                message += f" Use {replacement}() instead."
+            if reason:
+                message += f" {reason}"
+            warnings.warn(message, DeprecationWarning, stacklevel=2)
+            return func(*args, **kwargs)  # Pass all arguments properly
+
+        return wrapper
+
+    return decorator
+
+
 class GeneralExceptions:
     class ValidationError:
-        @custom_exception_str
-        class FileNotFound(FileNotFoundError):
-            def __init__(self, fp: str, **kwargs: Kwargs) -> None:
+        class FileNotFound(FileNotFoundError, CustomBaseException):
+            def __init__(self, fp: str) -> None:
                 """
                 Raised when a file in a given path is not found.
 
                 Args:
                 - parameter (`fp`): Path of the file that can not be found.
+
                 """
                 self.message = f"`{fp}` does not exist."
 
-        @custom_exception_str
-        class Arguments(Exception):
+        class Arguments(CustomBaseException):
             def __init__(
                 self,
                 parameter: str,
@@ -38,15 +77,15 @@ class GeneralExceptions:
                 - parameter (`str`): Name of the parameter.
                 - argument (`any`): Argument passed to the parameter.
                 - specification (`str`): Specification/s of the parameter.
+
                 """
                 self.message = f"Argument `{parameter}` needs to {specification}. Instead, passed in the following: {argument}"
 
         @custom_exception
-        class Common(Exception):
+        class Common(CustomBaseException):
             pass
 
-    @custom_exception_str
-    class PrerequisiteNotFound(Exception):
+    class PrerequisiteNotFound(CustomBaseException):
         def __init__(
             self,
             prerequisite: str,
@@ -59,70 +98,82 @@ class GeneralExceptions:
             Args:
             - prerequisite (`str`): Name of the prerequisite.
             - inst_instruction (`Optional[str]`, optional): Instructions for installing the prerequisite. Defaults to `None`.
+
             """
             self.message = f"prerequisite `{prerequisite}` cannot be found."
 
 
 class CLIExceptions:
-    @custom_exception_str
-    class TerminalTooThin(Exception):
-        def __init__(self, min_width: int, **kwargs: Kwargs) -> None:
+    class TerminalTooThin(CustomBaseException):
+        def __init__(self, min_width: int) -> None:
             """
             Raised when terminal is too thin for content to be rendered.
 
             Args:
             - min_width (`int`): Required minimum terminal width.
+
             """
             self.message = f"Please widen terminal.\nCurrent Width: {TW}\nMinimum Width: {min_width}"
 
     class ValidationError:
-        @custom_exception_str
-        class OptionRequired(Exception):
-            def __init__(self, option: str, **kwargs: Kwargs) -> None:
+        class OptionRequired(CustomBaseException):
+            def __init__(self, option: str) -> None:
                 """
                 Raised when an option is required but no argument is passed.
 
                 Args:
                 - option (`str`): Required option with no arguments passed into it.
+
                 """
                 self.message = f"Option `{option}` is required."
 
         @custom_exception
-        class Common(Exception):
+        class Common(CustomBaseException):
             pass
 
 
+class ConfigExceptions:
+    class ExtensionNotSupported(NotImplementedError, CustomBaseException):
+        def __init__(self, ext: str) -> None:
+            """
+            Raise when extension `{ext}` is not supported.
+
+            Args:
+                ext (`str`): The extension not supported.
+
+            """
+            self.message = f"Extension `{ext}` is not supported."
+
+
 class CFGExceptions:
-    @custom_exception_str
-    class ExtensionNotSupported(NotImplementedError):
-        def __init__(self, ext: str, **kwargs: Kwargs) -> None:
+    @deprecated(
+        "3.0.0",
+        "alltheutils.exceptions.ConfigExceptions.ExtensionNotSupported",
+    )
+    class ExtensionNotSupported(NotImplementedError, CustomBaseException):
+        def __init__(self, ext: str) -> None:
             self.message = f"Extension `{ext}` is not supported."
 
 
 class NestedDictExceptions:
-    @custom_exception_str
-    class NonDictReplacementValue(TypeError):
+    class NonDictReplacementValue(TypeError, CustomBaseException):
         def __init__(self) -> None:
             self.message = "Cannot replace a dict with a non-dict."
 
-    @custom_exception_str
-    class ValueNotAList(TypeError):
+    class ValueNotAList(TypeError, CustomBaseException):
         def __init__(self, keys: list[str], idx: int) -> None:
             self.message = f"Value of path '{'.'.join(keys[: idx + 1])}' is not a list."
 
-    @custom_exception_str
-    class ValueNotADict(TypeError):
+    class ValueNotADict(TypeError, CustomBaseException):
         def __init__(self, keys: list[str], idx: int) -> None:
             self.message = f"Value of path '{'.'.join(keys[:idx])}' is not a dict."
 
-    @custom_exception_str
-    class ValueDoesNotExist(KeyError):
+    class ValueDoesNotExist(KeyError, CustomBaseException):
         def __init__(self, keys: list[str], idx: int) -> None:
             self.message = (
                 f"Value of path '{'.'.join(keys[: idx + 1])}' does not exist."
             )
 
-    @custom_exception_str
-    class ValueIsAListAndIndexIsOutOfRange(IndexError):
+    class ValueIsAListAndIndexIsOutOfRange(IndexError, CustomBaseException):
         def __init__(self, keys: list[str], idx: int) -> None:
             self.message = f"Value of path '{'.'.join(keys[: idx + 1])}' is a list and index is out of range."