+++ alltheutils/utils.py
@@ -12,7 +12,7 @@ import sys
 import unicodedata
 import warnings
 from collections.abc import Callable, Generator, Iterable, Sized
-from datetime import datetime
+from datetime import datetime, timezone
 from itertools import cycle
 from multiprocessing import Pool, pool
 from os import makedirs
@@ -24,8 +24,6 @@ from typing import Any, Final, Optional
 from alltheutils import PSH, types
 from alltheutils.exceptions import GeneralExceptions, NestedDictExceptions
 
-warnings.simplefilter("always")
-
 # ================================ Constants ===================================
 PR = ["alpha", "beta", "rc"]  # Prerelease strings
 CATEGORIES: Final[set[str]] = {"Cn"}
@@ -33,9 +31,9 @@ CATEGORIES: Final[set[str]] = {"Cn"}
 # ============================ Derived Constants ===============================
 ALL_CHARS: Final[Generator[str, None, None]] = (chr(i) for i in range(sys.maxunicode))
 CCHARS: Final[str] = "".join(
-    map(chr, itertools.chain(range(0x00, 0x20), range(0x7F, 0xA0))),
+    map(chr, itertools.chain(range(0x20), range(0x7F, 0xA0))),
 )
-CCHARS_RE: Final[Pattern[str]] = re.compile("[%s]" % re.escape(CCHARS))  # type: ignore
+CCHARS_RE: Final[Pattern[str]] = re.compile("[{}]".format(re.escape(CCHARS)))  # type: ignore
 
 
 # ================================= Classes ====================================
@@ -73,8 +71,10 @@ def deprecated(
     Decorator to mark functions as deprecated.
 
     Args:
-        version (str): The version in which the function will be removed.
-        replacement (str, optional): The new function to use instead.
+    - version (`str`): The version in which the function will be removed.
+    - replacement (`str`, optional): The new function to use instead.
+    - reason (`str`, optional): Reason for deprecation. Defaults to None.
+
     """
 
     def decorator(func):
@@ -97,6 +97,11 @@ def deprecated(
     return decorator
 
 
+@deprecated(
+    "3.0.0",
+    "alltheutils.multi_processing.run_mp_qir",
+    "This is probably broken anyways.",
+)
 def run_mp_qir(
     func: types.CallableAny,
     iterable: types.IterAny,
@@ -108,6 +113,12 @@ def run_mp_qir(
     Iterate over `iterable` and apply iterated item to `func` asynchronously. Wait for a single process in the pool to return, and terminate the pool.
 
     This function requires the given function to return a bool, or an iterable with its first item as a bool. This bool is then used to decide whether to trigger the callback and terminate the pool.
+
+    Args:
+    - func (`types.CallableAny`): Function to be run on each item in parallel.
+    - iterable (`types.IterAny`): Iterable containing items to iterate over and pass to `func`.
+    - callback (`types.CallableAny`): Function to be called when a process in the pool returns.
+
     """
     if callback is None:
         callback = noop
@@ -122,6 +133,11 @@ def run_mp_qir(
         pool.join()
 
 
+@deprecated(
+    "3.0.0",
+    "alltheutils.multi_processing.run_mp_star_qir",
+    "This is probably broken anyways.",
+)
 def run_mp_star_qir(
     func: types.CallableAny,
     iterable: types.IterIterAny,
@@ -152,6 +168,7 @@ def batch_replace(text: str, key_value_map: dict[str, list[str]]) -> str:
 
     Returns:
     `str`: Modified string.
+
     """
     for k, v in key_value_map.items():
         if v:
@@ -165,10 +182,11 @@ def calculate_sha256_hash(input: str) -> str:
     Given a string, calculate its hash and return it.
 
     Args:
-        input (str): String to hash.
+    - input (str): String to hash.
 
     Returns:
-        str: Hash of the string.
+    `str`: Hash of the string.
+
     """
     sha256_hash = hashlib.sha256()
     sha256_hash.update(input.encode("utf-8"))
@@ -189,6 +207,7 @@ def caller_relative_path(relative_path: str, idx: Optional[int] = None) -> str:
 
     Returns:
     `str`: Path relative to the caller's path.
+
     """
     if idx is None:
         idx = 1
@@ -208,6 +227,7 @@ def custom_version_ls_to_str(
 
     Returns:
     `tuple[str, str]`: List of string representation of given list of version numbers, both in modified semver form and semver-compliant form.
+
     """
     pr = ""
     ivls = [int(i) for i in vls]
@@ -234,6 +254,7 @@ def dict_get_first_match(
 
     Returns:
     `Any`: Retrieved value.
+
     """
 
     for i in keys:
@@ -255,6 +276,7 @@ def ensure_parent_dir(
 
     Returns:
     `str`: Given filepath.
+
     """
 
     pd = os.path.dirname(file_path)
@@ -277,6 +299,7 @@ def file_exists(fp: str) -> str:
 
     Returns:
     `str`: Return `fp` when file path exists.
+
     """
     if not os.path.exists(fp):
         raise GeneralExceptions.ValidationError.FileNotFound(fp)
@@ -299,6 +322,7 @@ def fill_ls(
 
     Returns:
     `types.SequenceAny`: Filled list.
+
     """
     lls = len(ls)
     if lls < length:
@@ -318,6 +342,7 @@ def first_not_none_in_ls(ls: types.ListOptionalAny) -> Any:
 
     Returns:
     `Any`: The first argument that is not `None`, else `None`.
+
     """
     for i in ls:
         if i is not None:
@@ -439,6 +464,7 @@ def if_none(variable: Any, default: Any) -> Any:
 
     Returns:
     `Any`: `variable` if `variable` is not None else `default`.
+
     """
     if variable is None:
         return default
@@ -462,6 +488,7 @@ def literal_eval(expr: Optional[str]) -> Any:
 
     Returns:
     `Any`: Expression literally evaluated.
+
     """
     if expr is not None:
         return ast.literal_eval(expr)
@@ -487,6 +514,7 @@ def parent_dir_nth_times(filename: str, n: Optional[int] = None) -> str:
 
     Returns:
     `str`: Parent directory of the given filename.
+
     """
     op = filename
     for _ in range(n or 1):
@@ -500,26 +528,45 @@ def run_cmd(cmd: str) -> None:
 
     Args:
     - cmd (`str`): Shell command to excute.
+
     """
-    call(shlex.split(cmd))
+    call(shlex.split(cmd))  # noqa: S603
 
 
+@deprecated(
+    "3.0.0", "alltheutils.multi_processing.run_mp", "This is probably broken anyways.",
+)
 def run_mp(func: types.CallableAny, iterable: types.IterAny) -> types.ListAny:
     with Pool() as pool:
         return pool.map(func, iterable)
 
 
+@deprecated(
+    "3.0.0",
+    "alltheutils.multi_processing.run_mp_qgr",
+    "This is probably broken anyways.",
+)
 def run_mp_qgr(func: types.CallableAny, iterable: types.IterAny) -> types.TupleAny:
     res_cb = CallbackGetResult()
     run_mp_qir(func, iterable, res_cb.callback)
     return res_cb.get()
 
 
+@deprecated(
+    "3.0.0",
+    "alltheutils.multi_processing.run_mp_star",
+    "This is probably broken anyways.",
+)
 def run_mp_star(func: types.CallableAny, iterable: types.IterIterAny) -> types.ListAny:
     with Pool() as pool:
         return pool.starmap(func, iterable)
 
 
+@deprecated(
+    "3.0.0",
+    "alltheutils.multi_processing.run_mp_star_qgr",
+    "This is probably broken anyways.",
+)
 def run_mp_star_qgr(
     func: types.CallableAny,
     iterable: types.IterIterAny,
@@ -540,6 +587,7 @@ def sanitize_text(s: str) -> str:
 
     Returns:
     `str`: Sanitized text.
+
     """
 
     if isinstance(s, str):
@@ -564,6 +612,7 @@ def search_query(
 
     Returns:
     `Generator[tuple[None, str] | tuple[float, str], None, None]`: Generator object of mastching search quries.
+
     """
 
     sequence_matcher = difflib.SequenceMatcher()
@@ -587,10 +636,11 @@ def str2int(num: int | str) -> Optional[int]:
     If given number is int, return it. Else, if given number is string and is decimal, convert string to integer. Otherwise, return None.
 
     Args:
-        s (int | str): int or string to convert to integer.
+    - s (`int | str`): int or string to convert to integer.
 
     Returns:
-        Optional[int]: If given argument can be converted to integer, it will be returned. Otherwise, None will be.
+    `Optional[int]`: If given argument can be converted to integer, it will be returned. Otherwise, None will be.
+
     """
     if isinstance(num, int):
         return num
@@ -610,9 +660,10 @@ def unix_timestamp_to_iso(timestamp: int) -> str:
 
     Returns:
     `str`: Formatted datetime string
+
     """
 
-    return datetime.fromtimestamp(timestamp, tz=datetime.UTC).strftime(
+    return datetime.fromtimestamp(timestamp, tz=timezone.UTC).strftime(  # type: ignore
         "%Y-%m-%dT%H:%M:%S",
     )
 
@@ -626,11 +677,12 @@ def which_ls(  # noqa: C901
     Given a command, mode, and a PATH string, return the path which conforms to the given mode on the PATH, or None if there is no such file. Yoinked from shutil.
 
     Args:
-        mode (Optional[int], optional): File mode to look for. Defaults to `os.F_OK | os.X_OK`.
-        path (Optional[str], optional): Path to search the command at. Defaults to the result of os.environ.get("PATH").
+    - mode (`int`, optional): File mode to look for. Defaults to `os.F_OK | os.X_OK`.
+    - path (`str`, optional): Path to search the command at. Defaults to the result of os.environ.get("PATH").
 
     Returns:
-        Optional[types.TupleStr]: Tuple of commands that conforms to the given arguments as said above.
+    `Optional[types.TupleStr]`: Tuple of commands that conforms to the given arguments as said above.
+
     """
 
     if mode is None:
@@ -711,7 +763,8 @@ def zip_extend(a: Sized, b: Sized) -> Iterable[Any]:
     - b (`Sized`): Second sized iterable.
 
     Returns:
-    `Iterable[Any]`: _description_
+    `Iterable[Any]`: Zipped extended iterable.
+
     """
     if len(a) > len(b):
         return zip(a, cycle(b), strict=False)  # type: ignore[arg-type, call-overload, no-any-return]